Add AnimatedList & PlaceholderSwitcher helpers components - Added PlaceholderSwitcher to help replacing placeholders with data - Added AnimatedList to animate list with fancy trail - Cleanup EventList
